requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.1.1
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
contourpy==1.2.0
crontab==1.0.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
docker==7.0.0
dpath==2.1.6
dulwich==0.21.5
dvc==3.15.3
dvc-data==2.14.0
dvc-http==2.30.2
dvc-objects==1.0.0
dvc-render==0.5.3
dvc-studio-client==0.13.0
dvc-task==0.3.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.2
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.1
Flask-MySQLdb==2.0.0
Flask-RQ2==18.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.49.0
fqdn==1.5.1
freezegun==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.32
grandalf==0.8
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==7.0.1
incoming==0.3.1
ipykernel==6.25.2
ipython==8.16.0
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kombu==5.3.1
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mizani==0.11.1
mlflow==2.11.0
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.2.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
orjson==3.9.5
overrides==7.4.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
plotnine==0.13.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
pygit2==1.12.2
Pygments==2.16.1
pygtrie==2.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
redis==5.1.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rq==1.16.2
rq-scheduler==0.13.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-learn==1.4.1.post1
scipy==1.12.0
scmrepo==1.2.1
Send2Trash==1.8.2
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
sqltrie==0.7.0
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.3.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.6
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0
